drivers: i2c: mchp: Fix Timeout errors
When a slave stretches the clock beyond the permissible limit
set in the driver, a timeout occurs post which the master deems
the transaction a failure and aborts. The recovery from this state
was broken (master was holding the clock low forever post this)
and that has been fixed with this commit.

Signed-off-by: Rajavardhan Gundi <rajavardhan.gundi@intel.com>